auto-ack messages
* A deadlock can occur when an error happens between consuming a message
and acking a message. This problem is compounded by the fact that our
messages have no TTL and our QoS settings only allow for 1 messages to
be consumed at a time.
* This change fixes the deadlock by auto-acking messages that are
consumed. We are safe to do this because of the nature of websocket
messages, they are transient. They do not require guaranteed delivery.